requirements.txt
torch>=1.12.0
torchvision>=0.13.0
torch-geometric>=2.0.0
numpy>=1.21.0
networkx>=2.6.3
scipy>=1.7.0
matplotlib>=3.4.3
scikit-learn>=1.0.0
pyyaml>=5.4.0
tqdm>=4.62.0
pandas>=1.3.0
